config,memtier: move marshallable time.Duration to config.
